Delete packages folder, move all its code to apps/extension
