+++ apps/extension/package.json
@@ -22,10 +22,6 @@
     "tc": "tsc --noEmit"
   },
   "dependencies": {
-    "@project/constants": "*",
-    "@project/types": "*",
-    "@project/utils": "*",
-    "@project/validation": "*",
     "axios": "^1.6.2",
     "axios-rate-limit": "^1.3.0",
     "classnames": "^2.3.2",
@@ -45,6 +41,7 @@
     "@types/chrome": "^0.0.251",
     "@types/firefox-webext-browser": "^111.0.4",
     "@types/lodash": "^4.14.201",
+    "@types/node": "^20.9.2",
     "@types/react": "^18.2.37",
     "@types/react-dom": "^18.2.15",
     "@typescript-eslint/eslint-plugin": "5.41.0",
