+++ apps/extension/src/types/general.ts
