+++ apps/extension/src/types/index.ts
@@ -1,4 +1,6 @@
-export * from '@project/types';
+export type { NexusHub } from './nexushub';
+export * from './api';
+export * from './general';
 
 export * from './data';
 export * from './reactQuery';
