--- packages/constants/CHANGELOG.md
@@ -1,13 +1,0 @@
-# @project/constants
-
-## 0.0.2
-
-### Patch Changes
-
-- Update to push new version to stores
-
-## 0.0.1
-
-### Patch Changes
-
-- Update to WOTLK
