--- packages/constants/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@project/constants",
-  "version": "0.0.2",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "license": "MIT",
-  "scripts": {
-    "build": "cross-env NODE_ENV=production tsup index.ts --dts",
-    "dev": "cross-env NODE_ENV=development tsup index.ts --dts --watch"
-  },
-  "devDependencies": {
-    "@project/types": "*",
-    "@types/node": "^20.9.1",
-    "tsup": "5.12.6",
-    "typescript": "*"
-  }
-}
