--- packages/constants/tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "CommonJS",
-    "target": "ES6",
-    "lib": ["esnext", "DOM"],
-    "moduleResolution": "node",
-    "noEmit": true,
-    "removeComments": true,
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "strict": true,
-    "resolveJsonModule": true,
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
-    "exactOptionalPropertyTypes": true,
-    "noUncheckedIndexedAccess": true,
-    "strictPropertyInitialization": false
-  },
-  "exclude": ["node_modules"]
-}
