--- packages/types/index.ts
@@ -1,3 +1,0 @@
-export type { NexusHub } from './nexushub';
-export * from './api';
-export * from './general';
