--- packages/types/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "@project/types",
-  "version": "0.0.1",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "license": "MIT",
-  "scripts": {
-    "build": "tsup index.ts --dts",
-    "dev": "tsup index.ts --dts --watch"
-  },
-  "devDependencies": {
-    "tsup": "5.12.6",
-    "typescript": "*"
-  }
-}
