--- packages/types/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "noImplicitAny": true,
-    "strictNullChecks": true,
-    "outDir": "dist"
-  }
-}
