--- packages/utils/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@project/utils",
-  "version": "0.0.1",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "license": "MIT",
-  "scripts": {
-    "build": "tsup src/index.ts --dts",
-    "dev": "tsup src/index.ts --dts --watch"
-  },
-  "devDependencies": {
-    "@project/types": "*",
-    "tsup": "5.12.6",
-    "typescript": "*"
-  }
-}
