--- packages/validation/CHANGELOG.md
@@ -1,7 +1,0 @@
-# @project/validation
-
-## 0.0.1
-
-### Patch Changes
-
-- Update to push new version to stores
