--- packages/validation/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "@project/validation",
-  "version": "0.0.1",
-  "main": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "license": "MIT",
-  "scripts": {
-    "build": "tsup src/index.ts --dts",
-    "dev": "tsup src/index.ts --dts --watch"
-  },
-  "devDependencies": {
-    "@project/types": "*",
-    "tsup": "5.12.6",
-    "typescript": "*"
-  },
-  "dependencies": {
-    "class-transformer": "^0.5.1",
-    "class-validator": "^0.13.2"
-  }
-}
